# internal/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Substrate REST API
  version: v1
  description: |-
    One repository per token, one set of routes for every kind. A kind
    reference is `{authority}/{package}/{kind}`, and a record's REST path is
    `/api/v1/` followed by that reference and its id: three segments address
    a collection, four a record. A new kind adds no route.

    Authentication is a bearer token on every route under `/api/v1` and on
    `/tokens`: `Authorization: Bearer substrate_tok_...`. A token has full
    access to its repository; there are no scopes. The token names the
    repository, so no path carries one. `X-Substrate-Actor` names the door a
    write came through (`console`, `substratectl`; absent is `api`) and is
    attribution only; the substrate's own hands (`substrate`, `bundle:`,
    `function:`, `agent:` names) are refused with `403 forbidden`.

    Every refused request answers one body, `ErrorEnvelope`, whose `code` is
    one of thirteen strings; the status code follows the code. A `compacted`
    refusal (`410`) names the changelog `head` and `generation` to re-list
    from.

    Lists page with an opaque `cursor` passed back as `after`; the response
    carries `head` and `generation`, the changelog position the list was read
    at, which a watch resumes from (`from={head}&generation={generation}`).
    The change feed pages with real sequence numbers, `from` forward and
    `before` backward, each bound to a history `generation`.

...
    nothing.

...

...
servers:
...
security:
  - bearer: []
...
paths:
...
  /api/v1/graphql:
    post:
      tags: [graphql]
      summary: GraphQL (preview)
      description: |-
        The GraphQL door. PREVIEW: the schema is generated per repository
        from its kinds, and every part of it, the root operations and the
        scalars included, may change without a v1 wire break. The supported
        interface is the REST surface in the rest of this document. A
        resolver error carries the REST problem object under `extensions`.
      operationId: graphql
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GraphQLRequest"
      responses:
        "200":
          description: The GraphQL result. Field errors are inside the body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphQLResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: |-
        A token secret, `substrate_tok_` and 40 hex characters, minted by
        `POST /login`, `POST /register` or `POST /tokens`.
...
  responses:
    Error:
      description: |-
        A refused request. `code` says which of the thirteen refusals; the
        status follows it: `bad_request` 400, `auth` 401, `guard` and
        `forbidden` 403, `not_found` 404, `conflict` 409, `compacted` 410,
        `validation` 422, `rate_limited` 429, `internal` and
        `function_failed` 500, `unsupported` 501, `unavailable` 503 (always
        with `Retry-After`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      description: The one body every refused request answers with.
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorPayload"
    ErrorPayload:
      type: object
      description: |-
        The refusal: a code from the closed set, a message, and for a
        validation refusal the problems as strings and split. `head` and
        `generation` ride a `compacted` refusal only.
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - validation
            - conflict
            - guard
            - forbidden
            - auth
            - not_found
            - rate_limited
            - compacted
            - internal
            - function_failed
            - unsupported
            - unavailable
        message:
          type: string
        problems:
          type: array
          description: "`path: message` strings."
          items:
            type: string
        problemDetails:
          type: array
          items:
            $ref: "#/components/schemas/ProblemDetail"
        head:
          type: integer
          format: int64
          description: The changelog head to re-list from (`compacted` only).
        generation:
          type: string
          description: The history generation of `head` (`compacted` only).
    ProblemDetail:
      type: object
      description: One validation problem, field-addressable.
      required: [path, message]
      properties:
        path:
          type: string
          description: The input the problem concerns (`props.name`); empty
            when unaddressed.
        message:
          type: string
...
    GraphQLRequest:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query:
          type: string
        variables:
          type: object
          additionalProperties: true
        operationName:
          type: string
        extensions:
          type: object
          additionalProperties: true
    GraphQLResponse:
      type: object
      required: [data]
      properties:
        data: {}
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
        extensions:
          type: object
          additionalProperties: true